New Home Page. Changed permissions.
New home page is actually just a modification to the old one. Now displays a welcome message if you visit whilst not logged in (as either student or club)
Has links to log in and sign up.

Changed permissions so only clubs can create edit or destroy events.
Did some trickery to try and prevent students and clubs from being able to be logged in at the same time.